backend/requirements.txt
# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Database - DynamoDB
boto3>=1.35.0
# For local DynamoDB testing (optional):
# boto3-stubs[dynamodb]>=1.35.0


# Auth
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.20
requests>=2.31.0

# Utilities
aiofiles>=24.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Email (for notifications)
# resend (add when ready to implement)

# Development
pytest>=8.3.0
pytest-asyncio>=0.24.0
black>=24.10.0
ruff>=0.8.0
